backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.13.0
Pillow==11.0.0
PyJWT==2.1.0
requests==2.26.0
djoser==2.1.0
setuptools==75.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
drf-extra-fields==3.7.0